Ensure brew is initialized before anything else
I was having an error when opening up a new terminal window where the
starship command was not being found, even when it was installed. This
was due to brew not being initialized at the correct time. This caused
starship, which is installed by brew, to not be found.

Changing the order of the evals to have starship evaluated after the
brew shellenv fixed it.