Update configs and add Nvim submodule
I've added a gitignore file to help keep the files that should not be
commited out of this repo. This is done by using path negation in the
gitignore file as the files that should be committed are far fewer than
the generated files that come from package managers, scripts, and
applications.

I've also added nvim configuration as a submodule. The main reason for
this is to help me keep my neovim config as a separate repo for simpler
sharing.

The gcloud config folder will now only include a default configuration
that only configures the use of the gcloud emulators included with the
CLI.

The only change to the git configuration is the inclusion of the
signingkey fingerprint.