Complete refactor
This commit is the initial implementation of three new tools:
- gnu stow
- 1password-cli
- antidote.sh
GNU Stow is now managine the links of all config files. To enable this the
directory structure has been updated to follow the Stow naming conventions.
Antidote is now replacing antigen to manage ZSH plugins
1password-cli, along with 1Password, is now managing security and ssh keys.
This commit is just the initial changes. More need to be made to have tmux and
other tools work for my workflow.