Add Obsidian plugin
I've added the Obsidian plugin with it's default config. Important to
note that I'm using the fork and not the original repo. The main reason
for this is that this fork includes updates to have the plugin work with
snacks.nvim and blink.cmp.

This would ensure that the plugin works well with the latest Lazyvim and
probably other NVIM distros.