Add configuration file for Lazy
I've added a new config file to update the Lazy.nvim config options.
This commit adds an update to increase the autocheck for plugins to be
every day instead of every hour. The main reason for this is having to
authorize the use of the personal ssh key is getting really tiresome.